+++ src/commands/index.ts
@@ -1,0 +1,5 @@
+export * as api from './api';
+export * as buildContract from './buildContract';
+export * as installHeaders from './installHeaders';
+export * as scaffoldContract from './scaffoldContract';
+export * as shared from './shared';