+++ src/commands/shared.ts
@@ -1,0 +1,13 @@
+export const commandNames = {
+    installHeaders: 'ultra.installHeaders',
+    buildContract: 'ultra.buildContract',
+    scaffoldContract: 'ultra.scaffoldContract',
+    api: 'ultra.api',
+};
+
+export const installed: { [K in keyof typeof commandNames]: boolean } = {
+    buildContract: false,
+    installHeaders: false,
+    scaffoldContract: false,
+    api: false,
+};