+++ src/utility/sleep.ts
@@ -1,0 +1,5 @@
+export async function sleep(msTimeout: number) {
+    return new Promise((resolve: Function) => {
+        setTimeout(() => resolve(), msTimeout);
+    });
+}